+++ src/common-utils/_zz_args.sh
@@ -21,6 +21,12 @@ if test $# -lt 1; then
         <argname> <datatype> <varname> <help>
         ...
         help" >&2
+
+    echo "With <argname>:
+    x   for flags x with value (e.g., -f value, -h being reserved for help)
+    -   for sequential arguments in the order defined, without flags
+    +   to capture all remaining arguments as a single variable with spaces as separators
+    &   to capture all remaining arguments as a single variable with newlines as separators" >&2
     return 1
 fi
 
@@ -46,6 +52,9 @@ while read argname datatype varname help; do
         helpinfo="$helpinfo\n\t$(printf '%-12s : %s' "$name" "$help")"
     elif [ "$argname" = "+" ]; then
         line="<$datatype...>"
+        helpinfo="$helpinfo\n\t$(printf '%-12s : %s' "$name" "$help")"
+    elif [ "$argname" = "&" ]; then
+        line="<$datatype...>>"
         helpinfo="$helpinfo\n\t$(printf '%-12s : %s' "$name" "$help")"
     else
         if [ "$datatype" = "-" ]; then
@@ -108,17 +117,25 @@ else
         fi
     done
 
-    # Process remaining '=' parameters
-    for arg in $(echo $varnames | grep -E "^=" | cut -f2); do
-        if [ "$#" -gt "0" ]; then
-            echo "$arg='$1'" && shift 1
-        fi
+    # Process remaining '&' parameters
+    for arg in $(echo $varnames | grep -E "^&" | cut -f2); do
+        lines=""
+        while [ "$#" -gt "0" ]; do
+            # spaces that are not escaped should be preserved as part of the argument
+            if [ -z "$lines" ]; then
+                lines="$1"
+            else
+                lines="$lines\\\\n$1"
+            fi
+            shift 1
+        done
+        echo "$arg='$lines'"  
     done
 
     # Process remaining '+' parameters
     for arg in $(echo $varnames | grep -E "^\+" | cut -f2); do
         if [ "$#" -gt "0" ]; then
-            echo "$arg='$(echo $@ | sed "s/ /\t/g")'" && shift $#
+            echo "$arg='$@'" && shift $#
         fi
     done